* 'main' of https://github.com/Azure/azure-sdk-for-python: (1472 commits)
  update tests (#24716)
  Auto release debug (#24713)
  remove global config for multiapi (#24542)
  Update readme.python.md (#24710)
  [AutoRelease] t2-sqlvirtualmachine-2022-06-02-79158(Do not merge) (#24672)
  [AutoRelease] t2-batch-2022-06-01-88601(Do not merge) (#24659)
  [Doc] update doc for resolve_issues_effectively (#24642)
  temporarily skip aad tests (#24708)
  [formrecognizer] Adding 2022-06-30-preview work (#24701)
  Comment out assert for now, requires further investigation (#24691)
  Implement token autorefresh (#21834)
  [Identity] Update Service Fabric manual testing instructions (#21370)
  [dpg] add send request doc (#24700)
  [servicebus] missing docstring async ce (#24698)
  [KV] Log error when HSM security domain download fails (#24687)
  Parameterize Dockerfiles (#24637)
  update code owners (#24688)
  Conda release 2022 06 01 (#24635)
  Suppress supply chain warnings. (#24682)
  [translation] arch feedback (#24669)
  ...

+++ sdk/core/azure-core/azure/core/polling/_poller.py
@@ -257,12 +257,12 @@ class LROPoller(Generic[PollingReturnType]):
 
     @distributed_trace
     def wait(self, timeout=None):
-        # type: (Optional[int]) -> None
+        # type: (Optional[float]) -> None
         """Wait on the long running operation for a specified length
         of time. You can check if this call as ended with timeout with the
         "done()" method.
 
-        :param int timeout: Period of time to wait for the long running
+        :param float timeout: Period of time to wait for the long running
          operation to complete (in seconds).
         :raises ~azure.core.exceptions.HttpResponseError: Server problem with the query.
         """
